completed service and controller

+++ src/main/java/com/monocept/project/dto/ClaimResponseDTO.java
@@ -1,12 +1,15 @@
 package com.monocept.project.dto;
 
-import com.monocept.project.enums.ClaimStatus;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.NoArgsConstructor;
+
+import com.monocept.project.enums.ClaimStatus;
+
 import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 @Getter
 @Setter
@@ -18,7 +21,7 @@ public class ClaimResponseDTO {
     private String claimNumber;
     private String policyNumber;
     private String customerName;
-    private Double claimAmount;
+    private BigDecimal claimAmount;
     private String claimReason;
     private LocalDate incidentDate;
     private ClaimStatus claimStatus;
