+++ src/main/java/com/monocept/project/dto/PolicyPlanResponseDTO.java
@@ -1,11 +1,14 @@
 package com.monocept.project.dto;
 
+import java.math.BigDecimal;
+
 import com.monocept.project.enums.PremiumType;
 import com.monocept.project.enums.ProductType;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.NoArgsConstructor;
+
 import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 @Getter
 @Setter
@@ -17,8 +20,8 @@ public class PolicyPlanResponseDTO {
     private String productName;
     private ProductType productType;
     private String planName;
-    private Double coverageAmount;
-    private Double premiumAmount;
+    private BigDecimal coverageAmount;
+    private BigDecimal premiumAmount;
     private PremiumType premiumType;
     private Integer duration;
     private String termsAndConditions;
